Increase mob breeding limiter
The intention of this setting is to reduce lag caused by entity collision, but the current limit is way too low in terms of practical breeding. There would still be no more than 0.64 mobs standing on a single block on average after my edit, while the original limit was 0.2 mob per block. Considering the fact that cows, pigs and sheep all have a hitbox of 1x1 in area, this would still not introduce much entity collision while making it possible to breed a reasonable amount of cows/pigs for food.