Remove the height limit when players fly with an elytra
https://github.com/dniym/IllegalStack/wiki/PreventInfiniteElytraFlight
This glitch only affects versions above 1.13, and apparently the fix is that once the player reaches a certain altitude (around 300m with my testing), the player stops to deploy the elytra and therefore starts to freefall. Unless the elytra is deployed again, this may lead to potentially lethal falling damage.